requirements.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
comtypes==1.1.7
wikipedia==1.4.0
google==3.0.0
ecapture==0.1.8
 wolframalpha=4.0.0
twilio=6.46.0
python-decouple=3.3
